Modified packet layout and BT delay
As the title states, this commit just implements the new packet layout modification. It also changes the BT timeout to two seconds instead of one. The first part is to make packet communication easier while the second part is to accommodate for the long response time for each 20 bytes.

The next thing I need to do is implement the packet sender which will send a simple status message when an animal stands in front of the ultrasonic sensor for long enough. This sensor may eventually change to a button as well, but I'm not sure. The packet sender will also send a packet when the food level is low.

This functionality should be really simple to implement, let's hope...