Nvm, its working
Got the class system working and it works really well. Definitely overkill, but I got a lot of practice with classes and struts so it works out. Plus, the code it much more organized and easier to work with. This should make the other packets much easier to implement.

Next, I need to implement the give_food packet.