Mostly finished givefood and foodlow packets
This commit includes mostly finished version of both the givefood packet as well as the food_low packet. I implemented a pretty good send_packet system so adding the check_if_animal function should be very easy to add. All I need to do is check if a button has been pressed for an extended period of time, and then send a packet saying so.

This commit also includes a lot of TODO comments since I have a lot I need to fix up and adjust eventually. Most of it relates to testing and ideas so I won't worry about them too much. And the rest should be solved once I add the other functionality.

The most complicated part I think that is coming up will be the red_led delay function as well as the overall delay function. I need to add some functionality that prevents constant packet sending when an animal just sits in front of the feeder. This will most likely include a delay of some sort, where it will wait a few minutes or so before checking if an animal is in front of the feeder again. I don't think this will take too long, but it might be annoying to implement well.

I also finally figured out how to work with extern variables better, so I was able to make the bt_serial and sonar objects global. Along with that, the ultrasonic sensor will need to be adjusted for the actual size of the feeder. This is expected though so a constant is defined in the common header to help adjust this.

Overall, I made a decent amount of progress and should finish quite a lot more later today. I will also have to adjust some stuff on the Android side to work with this new set up. Also, I will need to start working on sending actual packets to the server, which is what I'm most worried about.