Animal packet added
This commit adds the animal packet and fixes some pin assignment issues. At this point, all of the basic packets are added but communicate with Bluetooth. Now I need to change the communication over to Wi-Fi for all packets other than the connect packet. This also includes making the PHP files for the packets as well as modifying and adding Bluetooth code for the Android app.