Working Bluetooth backup
This commit is at a point where everything is working by using Bluetooth. The android app is communicating and so is the website. I will be pushing each part of the code at this point to sync everything up.

After this, I am going to remove the ultrasonic code which is used to detect low food levels. We have decided to remove this since it shouldn't ever come to a point where you are low on food. If you do get this notification, you aren't checking in on your animal enough.

We have also decided to stick with a button for the animal detection instead of an ultrasonic sensor. This is due to the inconsistencies that come from the device, especially on a padded surface such as a dog's hair. We would eventually change this to a RFID tag, but that's not going to be implemented in this prototype.

Next, I need to change the Bluetooth communication over to Wi-Fi communication.