Overhaul for packet handling, changed naming scheme, and added mac address
This commit started with me struggling to get the full MAC address to be sent over. Because of that, I changed how I received data and simplified it dramatically.

Along with this, I modified some of the naming schemes and packet layout to make things more consistent.

Next, I need to send the animal packet to the webserver and compare the time to the feeders schedule. Hence the reason why I am saving the MAC address.