Mostly working project for presentation
Referencing my commit for the Android side of things, this one follows suit. This commit has all of the code for the project to function, albeit a few bugs, a lot of messy code, and some oversights. But overall, I'm happy with where the code is for now and it should work fine.

I will have to continue to clean this code up and organize it a bit more. I also over engineered it as usual, so things could be simplified quite a lot, unless I want to add more. Where it is right now, adding and removing packets is pretty easy. You just need to add the code for the functionality.

Off the top of my head, I can't really think of anything that needs to be addressed. I know it's not perfect, but until I use it a bit more, I don't see any major issues. Hopefully nothing will break during the presentation... Oh, actually, time zone things are an issue. I don't recall if I mentioned it in the Android commit, but I need to find a way to reset each feeder at midnight relative to that feeders time zone. This means I will probably have to mess around with packets a bit more and get some information from the internet for each feeder. It's probably going to be a pain anyway I do it.